server/pyproject.toml
[tool.poetry]
name = "or-server"
version = "0.1.0"
description = "Flask server for orchard robotics take home"
authors = ["Ross <rssmtdd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
flask-cors = "^4.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
